switch saving points in float & UI improvements

+++ Student.h
@@ -12,30 +12,30 @@ class Student{
 		string Fullname;
 		Date DoB;
 		int Gender; // 0 Male 1 Female 2 Other
-		int Math;
-		int Phys;
-		int Chem;
+		float Math;
+		float Phys;
+		float Chem;
 	public:
 		Student();
-		Student(int Id, string Fullname, Date DoB, int Gender, int Math = 0, int Phys = 0, int Chem = 0);
+		Student(int Id, string Fullname, Date DoB, int Gender, float Math = 0.0, float Phys = 0.0, float Chem = 0.0);
 		
 		int getId();
 		string getFullname();
 		Date getDoB();
-		int getMath();
-		int getPhys();
-		int getChem();
+		float getMath();
+		float getPhys();
+		float getChem();
 		int getGender();
 		
 		Student* setId(int Id);
 		Student* setFullname(string Fullname);
 		Student* setDoB(Date DoB);
-		Student* setMath(int Math);
-		Student* setPhys(int Phys);
-		Student* setChem(int Chem);
+		Student* setMath(float Math);
+		Student* setPhys(float Phys);
+		Student* setChem(float Chem);
 		Student* setGender(int Gender);
 		
-		int getAvg();
+		float getAvg();
 };
 
 class StudentNode{
